made LinkedList homework generic. moved solution files
