+++ day02/src/MyLinkedListTest.java
@@ -14,11 +14,7 @@ public class MyLinkedListTest {
 	@Before
 	public void setUp() throws Exception{
 
-		// QUESTIONS 1-7. COMMENT THIS LINE FOR QUESTION 7
-		chickens = new MyLinkedList();
-
-		// QUESTION 8. COMMENT THIS LINE FOR QUESTIONS 1-6
-		// chickens = new MyLinkedList<Chicken>();
+		chickens = new MyLinkedList<Chicken>();
 
 		chickens.add(new Chicken("Delilah", 10, "blue"));
 		chickens.add(new Chicken("Georgy", 8, "green"));
